feat: implemented `ProductCreationAIPromptProgressBar` inside the Starting Info View of Product Creation with AI
